docs: update READMEs to clarify reference to SLSA generators and builders
Previously, at some points accross the different READMEs there were
explanations on why it's necessary to refer the SLSA generators and
builders using tags in the format @vX.Y.Z, but some of them were
incomplete. This change adds a subtopic to the base README that
contains all the explanation, and this subtopic is linked on other
parts of READMEs that needs the explanation.

Signed-off-by: Diogo Teles Sant'Anna <diogoteles08@gmail.com>