feat: Add slsa-version input to setup-token (#1793)
Adds a `slsa-version` input to `setup-token`. Once we agree on this
approach we can add tests. The initial version will support v0.2
provenance but we can potentially remove before the final release of
BYOB.

In the future, I think we will still want to support an input that
allows users to generate different versions of SLSA provenance in order
to better support downstream consumers. For example, even if 1.1 is
released and we generate that by default users may want to generate 1.0
provenance because the downstream tooling hasn't caught up to 1.1 yet.
We can do this via the `slsa-version` input.

---------

Signed-off-by: Ian Lewis <ianlewis@google.com>