feat: Use sha512 as the digest algorithm in nodejs builder (#1779)
Fixes #1769

Uses the algorithm and digest from the npm integrity digest that is
outputted by `npm pack`. This is the same format as the integrity hashes
used by `package-lock.json`. However, these are encoded in base64 rather
than hex so we need to convert to hex in order to use it in the
attestation.

---------

Signed-off-by: Ian Lewis <ianlewis@google.com>