fix: Dont validate or read event path from slsa token (#1913)
Fixes #1900 

Adds a new `event_payload_sha256` field to the BYOB SLSA token. This
contains a sha256 checksum in hex format of the event payload as read
from `GITHUB_EVENT_PATH`. The checksum is compared against the checksum
of the same file when read by `verify-token`. This ensures that the same
event payload for the user's workflow is being included in provenance.

A new function `safeFileSha256` was added to `tscommon` since it's used
in a number of places. `tscommon` was updated to `0.1.0` and a new
tarball was added. Adding a new tarball will help with understanding
which actions depend on which version of `tscommon` without necessarily
having to update every action whenever `tscommon` changes. We can remove
old version of `tscommon` when we are sure they are not used.

---------

Signed-off-by: Ian Lewis <ianlewis@google.com>