fix: resolve target edge cases in Bazel Builder with cquery (#2324)
cc: @mihaimaruseac 

Changed getting the targets from parsing in shell script to using
Bazel's built in `cquery` command.

This resolves edge cases that resulted from the parsing schema.
Particularly, more generic glob patterns are supported such as
`//src/...` and `//src/internal:all`. Additionally, using cquery to
obtain target paths allows the user to have more flexibility in the
flags that they can use to configure the build. Users that pass the flag
to optimize the build process or users on different types of machines
will have a different path to the artifacts within the bazel-out
directory. Using cquery, gives the path to artifacts solving this
problem.

However, cquery does not have runfile support as it is based on the
build graph. Runfile support for artifacts is currently being developed.

---------

Signed-off-by: Noah Elzner <78953604+enteraga6@users.noreply.github.com>