feat: Add Maven BYOB publisher (#2242)
Adds a publisher for the Maven builder.

It is missing a few bits here and there, but I am keeping it open as a
draft to track progress.

The main things to fix are:

1. ~~Keep the built artifacts without modifying the delegator.~~ Done
2. ~~Use the safe downloaders and uploaders.~~ Done

We have made a Maven plugin to replace `collect_release_artifacts.sh`.
This plugin currently exist in the test projects source tree but should
ideally exist in Maven Central.

Example usage:
https://github.com/AdamKorcz/test-java-project/blob/main/.github/workflows/byob.yml
Example action:
https://github.com/AdamKorcz/test-java-project/actions/runs/5380001938

Sample release: Browse v0.1.17 here:
https://central.sonatype.com/artifact/io.github.adamkorcz/test-java-project/0.1.17/versions

---------

Signed-off-by: AdamKorcz <44787359+AdamKorcz@users.noreply.github.com>
Signed-off-by: AdamKorcz <adam@adalogics.com>
Signed-off-by: Noah Elzner <elzner@google.com>