feat: Runfile and java artifact support for Bazel Builder (#2362)
cc: @mihaimaruseac 

This will provide support for artifacts that also need runfiles. Users
can use a flag, `needs-runfiles`, to also package the artifact alongside
the runfiles Bazel generates if the user deems it as necessary.

Also, support for Java artifacts is made available by packaging the JARs
to be standalone through a format in Bazel called `_deploy.jar`.
Alongside the `_deploy.jar`, there will be a modified `run-script` that
allows the users that download the artifact to run the run-script by
using a flag called `local_javabin`, where they put the path to their
own java bin such that it is utilized by the run-script. This run-script
is generated by Bazel from a template and later modified in the
`build.sh` in the internal part of the builder to add this flag for the
users. More information is available on the readme.

Java targets will automatically be converted to their `_deploy.jar` with
this. Three flags are used for users that have java targets:

`includes-java`: if true then adds an additional flag to build command
as well as rule for local java repo in WORKSPACE in order to utilize the
`--singlejar` capability of run-script for `_deploy.jar` such that the
remote jdk does not need to be included in runfiles. Doing it like this
prevents massive bloat when attesting.

`user-java-distribution` and `user-java-version`: let the user specify
the exact java they want to use to build

When users run the run-script they will include an additional flag
`local_javabin` which they will set to their local javabin that the
run-script will utilize to run the `_deploy.jar`

A combination of `bazel query` and `bazel cquery` were used to resolved
edge cases with the implementation.

---------

Signed-off-by: Noah Elzner <78953604+enteraga6@users.noreply.github.com>
Signed-off-by: Noah Elzner <elzner@google.com>
Signed-off-by: Ian Lewis <ianlewis@google.com>
Signed-off-by: laurentsimon <laurentsimon@google.com>
Signed-off-by: laurentsimon <64505099+laurentsimon@users.noreply.github.com>
Co-authored-by: Ian Lewis <ianlewis@google.com>
Co-authored-by: laurentsimon <64505099+laurentsimon@users.noreply.github.com>