fix: fail if large file sha256 does not match (#2433)
As part of
https://github.com/slsa-framework/slsa-github-generator/pull/2365, we
added support for large files.

We forgot to update the "success" computation. This means that even
though the sha256 may differ from the expected, the job continues and we
never record the failure status.

I fix this in the PR. I also delete the downloaded folder / artifact if
the sha256 don't match, to avoid a similar problem in he future

---------

Signed-off-by: laurentsimon <laurentsimon@google.com>
Signed-off-by: Noah Elzner <elzner@google.com>