feat: add maven verification plugin (#2380)
Adds a Maven plugin that makes it easy to verify the provenance files of
all dependencies of a Maven-based project.

The plugin can either run during the Maven build cycle, or users can
invoke it manually from within a project directory.

The plugin checks whether the dependencies in a `pom.xml` file are
released with provenance attestations, and invokes the
[slsa-verifier](https://github.com/slsa-framework/slsa-verifier) against
the dependencies that have provenance files.

@loosebazooka @laurentsimon

---------

Signed-off-by: AdamKorcz <adam@adalogics.com>
Signed-off-by: AdamKorcz <44787359+AdamKorcz@users.noreply.github.com>
Co-authored-by: Ian Lewis <ianlewis@google.com>
Signed-off-by: Noah Elzner <elzner@google.com>