fix: Update maven helper plugin build (#3746)
# Summary

Attempt to fix
https://github.com/slsa-framework/slsa-github-generator/issues/3663#issuecomment-2126574002

Update the build definition for the maven slsa-hashing-plugin. It seems
to be pretty old and slightly misconfigured. @ianlewis @ramonpetgrave64
This might fix the testing issue? I'm not exactly sure how to trigger
the original failure though.

If after merging the fix doesn't work, then we will revert.

## Testing Process

We were able to partially test the fix locally, but we can't be 100%
sure until after merging.
We have an new pre-submit e2e test workflow that we can invoke from
within this repo to test the changes, though a more full test should
still be invoked from example-package.

## Checklist

- [x] Review the contributing
[guidelines](https://github.com/slsa-framework/slsa-github-generator/blob/main/CONTRIBUTING.md)
- [x] Add a reference to related issues in the PR description.
- [x] Update documentation if applicable.
- [x] Add unit tests if applicable.
- [ ] Add changes to the
[CHANGELOG](https://github.com/slsa-framework/slsa-github-generator/blob/main/CHANGELOG.md)
if applicable.

Signed-off-by: Appu Goundan <appu@google.com>
Signed-off-by: Ramon Petgrave <ramon.petgrave64@gmail.com>