chore: update refs to v2.1.0-rc.1 (#4120)
#label:release v2.1.0-rc.1

# Summary

Retrying the release, under a new tag, not rc.0.

The [original release
steps](https://github.com/slsa-framework/slsa-github-generator/blob/v2.0.0/RELEASE.md#verify-rc-version-references)
required doing a force push to update the tag in order to trigger
workflow to produce the binaries as release artifacts. But now, repo
settings make tags immutable, and I don't have permission to change the
settings, so we'll try something different:

1. Use the changes in this PR's branch (not main) to cut a new
pre-release under a new tag. This should produce the binaries and
associate it with the release.
-
https://github.com/slsa-framework/slsa-github-generator/releases/tag/v2.1.0-rc.1
1. Merge this PR, after doing step 1.
1. The release and main should now be at the same commit sha

## Testing Process

presubmits should pass

## Checklist

- [x] Review the contributing
[guidelines](https://github.com/slsa-framework/slsa-github-generator/blob/main/CONTRIBUTING.md)
- [x] Add a reference to related issues in the PR description.
- [x] Update documentation if applicable.
- [x] Add unit tests if applicable.
- [x] Add changes to the
[CHANGELOG](https://github.com/slsa-framework/slsa-github-generator/blob/main/CHANGELOG.md)
if applicable.

Signed-off-by: Ramon Petgrave <ramon.petgrave64@gmail.com>